Fixed product location and replaced with correct model file in the correct package
