+++ src/lk/supramart/dao/ProductDAO.java
@@ -1,6 +1,7 @@
 package lk.supramart.dao;
 
 import java.util.List;
+import lk.supramart.model.Product;
 
 public interface ProductDAO {
